Small mediapicker fixes (#3980)
* Small mediapicker fixes

* Shw only filename for selected files

* Hide crop modal after cropping

* Add condition

* Improve condition

* Improve multi/single decision

+++ src/Http/Controllers/Controller.php
@@ -141,11 +141,11 @@ abstract class Controller extends BaseController
                 $data->{$row->field} = str_replace($uuid, $data->getKey(), $data->{$row->field});
             });
             $data->save();
-
-            Storage::disk(config('voyager.storage.disk'))->move($old_path, $new_path);
-            Storage::disk(config('voyager.storage.disk'))->deleteDirectory($folder_path);
-
-            $request->session()->forget([$slug.'_path', $slug.'_uuid']);
+            if ($old_path != $new_path && !Storage::disk(config('voyager.storage.disk'))->exists($new_path)) {
+                $request->session()->forget([$slug.'_path', $slug.'_uuid']);
+                Storage::disk(config('voyager.storage.disk'))->move($old_path, $new_path);
+                Storage::disk(config('voyager.storage.disk'))->deleteDirectory($folder_path);
+            }
         }
 
         return $data;
